update unit tests for anta.cli.nrfu
