+++ anta/cli/nrfu/commands.py
@@ -6,17 +6,14 @@ Commands for Anta CLI to run nrfu commands.
 """
 from __future__ import annotations
 
-import asyncio
 import logging
 import pathlib
 
 import click
 
 from anta.cli.utils import exit_with_code
-from anta.models import AntaTest
-from anta.runner import main
 
-from .utils import anta_progress_bar, print_jinja, print_json, print_settings, print_table, print_text
+from .utils import anta_progress_bar, print_jinja, print_json, print_table, print_text
 
 logger = logging.getLogger(__name__)
 
@@ -30,9 +27,6 @@ logger = logging.getLogger(__name__)
 )
 def table(ctx: click.Context, device: str | None, test: str | None, group_by: str) -> None:
     """ANTA command to check network states with table result"""
-    print_settings(ctx)
-    with anta_progress_bar() as AntaTest.progress:
-        asyncio.run(main(ctx.obj["result_manager"], ctx.obj["inventory"], ctx.obj["catalog"], tags=ctx.params.get("tags")))
     print_table(results=ctx.obj["result_manager"], device=device, group_by=group_by, test=test)
     exit_with_code(ctx)
 
@@ -49,9 +43,6 @@ def table(ctx: click.Context, device: str | None, test: str | None, group_by: st
 )
 def json(ctx: click.Context, output: pathlib.Path | None) -> None:
     """ANTA command to check network state with JSON result"""
-    print_settings(ctx)
-    with anta_progress_bar() as AntaTest.progress:
-        asyncio.run(main(ctx.obj["result_manager"], ctx.obj["inventory"], ctx.obj["catalog"], tags=ctx.params.get("tags")))
     print_json(results=ctx.obj["result_manager"], output=output)
     exit_with_code(ctx)
 
@@ -62,9 +53,6 @@ def json(ctx: click.Context, output: pathlib.Path | None) -> None:
 @click.option("--skip-error", help="Hide tests in errors due to connectivity issue", default=False, is_flag=True, show_default=True, required=False)
 def text(ctx: click.Context, search: str | None, skip_error: bool) -> None:
     """ANTA command to check network states with text result"""
-    print_settings(ctx)
-    with anta_progress_bar() as AntaTest.progress:
-        asyncio.run(main(ctx.obj["result_manager"], ctx.obj["inventory"], ctx.obj["catalog"], tags=ctx.params.get("tags")))
     print_text(results=ctx.obj["result_manager"], search=search, skip_error=skip_error)
     exit_with_code(ctx)
 
@@ -89,8 +77,5 @@ def text(ctx: click.Context, search: str | None, skip_error: bool) -> None:
 )
 def tpl_report(ctx: click.Context, template: pathlib.Path, output: pathlib.Path | None) -> None:
     """ANTA command to check network state with templated report"""
-    print_settings(ctx, template, output)
-    with anta_progress_bar() as AntaTest.progress:
-        asyncio.run(main(ctx.obj["result_manager"], ctx.obj["inventory"], ctx.obj["catalog"], tags=ctx.params.get("tags")))
     print_jinja(results=ctx.obj["result_manager"], template=template, output=output)
     exit_with_code(ctx)
