+++ tests/data/template.j2
@@ -1,0 +1,3 @@
+{% for d in data %}
+* {{ d.test }} is [green]{{ d.result | upper}}[/green] for {{ d.name }}
+{% endfor %}
