--- tests/units/cli/nrfu/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) 2023 Arista Networks, Inc.
-# Use of this source code is governed by the Apache License 2.0
-# that can be found in the LICENSE file.
